fix : 코드 정리
