+++ Kung/Assets/Animation/PlayerColtroller/PlayerAniController/BodyAniController.controller
@@ -14,7 +14,7 @@ AnimatorStateMachine:
     m_Position: {x: 310, y: 120, z: 0}
   - serializedVersion: 1
     m_State: {fileID: 8795916396696396804}
-    m_Position: {x: 330, y: 240, z: 0}
+    m_Position: {x: 310, y: 210, z: 0}
   m_ChildStateMachines: []
   m_AnyStateTransitions: []
   m_EntryTransitions: []
