--- Kung/Assets/Scripts/Booster.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5bb0155f92cdb9d45a1714730adfba4a
