--- Kung/Assets/Scripts/CustomOreTile.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d738ba2a13e59224cb23fed7747f6ee4
