--- Kung/Assets/Scripts/Direction.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2260195a03c795a46bf8a92eb60aef19
