--- Kung/Assets/Scripts/Drilling.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 916f99d1c07e9f9449a0ca3b26b8fe03
