--- Kung/Assets/Scripts/InventoryUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e04351b8dcf9eef4496070623c6b6ec9
