--- Kung/Assets/Scripts/MineralDatabase.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d0c649ee188c68e4bbed83dd91600b7e
