--- Kung/Assets/Scripts/PlayerAniController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c770f87a70cf39247a34d37a436b2f57
