--- Kung/Assets/Scripts/PlayerHealth.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c571211c408cd4a4e8c057e6591c1203
