--- Kung/Assets/Scripts/PlayerMovement.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0d68d7e474324234ebc4f09d98554cb1
