--- Kung/Assets/Scripts/Shop.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a3b36c36113f96e41b01caf087e9e6a6
