+++ Kung/Assets/Scripts/Tile.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a0ed4cd6f2745b043a97a6794426ffc1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
