+++ Kung/Assets/Scripts/Tile/CustomOreTile.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d738ba2a13e59224cb23fed7747f6ee4
