+++ Kung/Assets/Scripts/UI/HUD.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 24aa12e3dceb4cf4487d9b1a34fbf356
